requirements.txt
streamlit>=1.31.0
librosa>=0.10.1
openai-whisper>=20231117
py-webrtcvad>=2.0.10
transformers>=4.37.2
torch>=2.2.0
numpy>=1.24.0
pandas>=2.1.0
soundfile>=0.12.1
python-dotenv>=1.0.0
openai>=1.12.0
SpeechBrain>=1.0.0 